subgraph: reorganize mappings into utils folder
