+++ packages/subgraph/src/mappings/utils/constants.ts
