+++ packages/subgraph/src/mappings/utils/helpers.ts
