+++ packages/subgraph/src/mappings/utils/sprintf.ts
